reorganize, remove unused files
