--- frontend/src/App/Settings/Days.js
@@ -1,9 +1,0 @@
-export const days = [
-    { name: 'Monday', number: 0 },
-    { name: 'Tuesday', number: 1 },
-    { name: 'Wednesday', number: 2 },
-    { name: 'Thursday', number: 3 },
-    { name: 'Friday', number: 4 },
-    { name: 'Saturday', number: 5 },
-    { name: 'Sunday', number: 6 }
-]
