+++ frontend/src/actions.js
@@ -1,5 +1,5 @@
 /* test or production servers */
-const domain = process.env.NODE_ENV === 'development' ? 
+export const domain = process.env.NODE_ENV === 'development' ? 
     'http://localhost:8072' : 'https://trans-bdit.intra.prod-toronto.ca/tt-request-backend'
 
 /* GET up to ten closest nodes given coordinate */
