--- frontend/src/colors.js
@@ -1,8 +1,0 @@
-export function getRandomColor(){
-    let letters = '0123456789ABCDEF';
-    let color = '#';
-    for (let i = 0; i < 6; i++) {
-        color += letters[Math.floor(Math.random() * 16)];
-    }
-    return color;
-}
